Spawner script
I tried to combine 2 random spawning scripts (one that spawns at random locations, and one that pick a random object on an array to spawn). The Random Object spawner part of the script seems to work, but I didn't have a chance to test it. The other half is set to inactive.